refactor: move pages components to the components folder
