+++ src/app/(home)/layout.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode } from 'react';
 
-import { Home } from '@/pages/HomePage/Home';
+import { Home } from '@/components/Home/Home';
 
 export default function MainLayout({ children }: { children: ReactNode }) {
   return <Home>{children}</Home>;
