+++ src/app/about/page.tsx
@@ -1,4 +1,4 @@
-import { About } from '@/pages/AboutPage/About';
+import { About } from '@/components/About/About';
 
 export default function aboutPage() {
   return <About />;
