+++ src/app/not-found.tsx
@@ -1,0 +1,5 @@
+import { NotFound } from '@/components/NotFound/NotFound';
+
+export default function NotFoundPage() {
+  return <NotFound />;
+}
