+++ src/components/Home/Home.tsx
@@ -1,26 +1,30 @@
 'use client';
 
-import type { PropsWithChildren } from 'react';
+import { type PropsWithChildren, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 
 import { FlyoutList } from '@/components/FlyoutList/FlyoutList';
 import { SearchBar } from '@/components/SearchBar/SearchBar';
-import { isClient } from '@/constants/common';
+import { useLocalStorage } from '@/hooks/useLocalStorage';
 import { useQueryParams } from '@/hooks/useQueryParams';
 import { checkedItemsSelectors } from '@/store/slices/checkedItems/selectors';
-import { SEARCH_TERM_STORAGE_KEY, setToLocalStorage } from '@/utils/localStorageUtils';
 
 import styles from './Home.module.scss';
 
 export function Home({ children }: PropsWithChildren) {
   const { appQueryParams, setQueryParams } = useQueryParams();
+  const [savedTerm, saveTerm] = useLocalStorage('task-anime', '');
   const hasSelected = !!useSelector(checkedItemsSelectors.selectTotal);
 
-  const handleSearchTermUpdate = (term: string) => {
-    if (isClient) {
-      setToLocalStorage(SEARCH_TERM_STORAGE_KEY, term);
-      setQueryParams({ details: undefined, page: undefined, query: term });
+  useEffect(() => {
+    if (savedTerm && !appQueryParams.query) {
+      setQueryParams({ query: savedTerm });
     }
+  }, [appQueryParams.query, savedTerm, setQueryParams]);
+
+  const handleSearchTermUpdate = (term: string) => {
+    saveTerm(term);
+    setQueryParams({ details: undefined, page: undefined, query: term });
   };
 
   return (
