--- src/pages/NotFoundPage/NotFound.module.scss
@@ -1,17 +1,0 @@
-@use '../../styles/placeholders' as *;
-@use '../../styles/mixins' as *;
-
-.wrapper {
-  @include flex(column, var(--spacing-3), $justify: center, $align: center);
-
-  max-width: 600px;
-  min-height: 80%;
-  margin: auto;
-  padding: var(--spacing-3);
-
-  text-align: center;
-
-  & button {
-    margin-top: var(--spacing-2);
-  }
-}
